+++ MyChat/Wiring/ExporterConfiguration.cs
