+++ MyChat/Writing/ConversationLog.cs
@@ -1,6 +1,10 @@
 namespace MindLink.Recruitment.MyChat
 {
     using System.Collections.Generic;
+    using System;
+    using System.IO;
+    using System.Security;
+    using Newtonsoft.Json;
 
     /// <summary>
     /// model for the output, to allow report to be optional in children
@@ -20,6 +24,34 @@ namespace MindLink.Recruitment.MyChat
         {
             this.name = conversation.name;
             this.messages = conversation.messages;
+        }
+
+        public void WriteLogToOutput(string outputFilePath)
+        {
+             try
+            {
+                var writer = new StreamWriter(new FileStream(outputFilePath, FileMode.Create, FileAccess.ReadWrite));
+
+                var serialized = JsonConvert.SerializeObject(this, Formatting.Indented);
+
+                writer.Write(serialized);
+
+                writer.Flush();
+
+                writer.Close();
+            }
+            catch (SecurityException)
+            {
+                throw new ArgumentException("No permission to file.");
+            }
+            catch (DirectoryNotFoundException)
+            {
+                throw new ArgumentException("Path invalid.");
+            }
+            catch (IOException)
+            {
+                throw new Exception("Something went wrong in the IO.");
+            }
         }
     }
